+++ tests/suites/message/PingTest.php
@@ -48,4 +48,14 @@ class PingTest extends TestCase
         $this->assertCount(1, $frames);
         $this->assertContainsOnlyInstancesOf(Frame::class, $frames);
     }
+
+    public function testPingPayload(): void
+    {
+        $message = new Ping('Some content');
+        $payload = $message->getPayload();
+        $this->assertEquals('U29tZSBjb250ZW50', base64_encode($payload));
+        $message = new Ping();
+        $message->setPayload($payload);
+        $this->assertEquals('Some content', $message->getContent());
+    }
 }